Move game queue to separate module and place it under the match runner #77
