+++ src/html/stats.ts
@@ -2,7 +2,7 @@ import {Config} from '../config';
 import * as cst from '../constants';
 import {AllImages} from '../imageloader';
 
-import {Game, schema} from 'battlecode-playback';
+import {schema} from 'battlecode-playback';
 
 const hex: Object = {
   1: "#a62014",
@@ -22,19 +22,9 @@ export default class Stats {
   readonly div: HTMLDivElement;
   private readonly images: AllImages;
 
-  // Match Options
-  private matches: HTMLDivElement;
-
   // Key is the team ID, folllowed by the robot/stat type
   private robotTds: Object = {};
   private statTds: Object = {};
-
-  // Callbacks initialized from outside Stats
-  // Yeah, it's pretty gross :/
-  onNextMatch: () => void;
-  onPreviousMatch: () => void;
-  gotoMatch: (game: number, match: number) => void;
-  removeGame: (game: number) => void;
 
   // Note: robot types and number of teams are currently fixed regardless of
   // match info. Keep in mind if we ever change these, or implement this less
@@ -49,7 +39,6 @@ export default class Stats {
     this.images = images;
 
     this.div = document.createElement("div");
-    this.matches = this.matchViewer();
 
     let teamNames: Array<string> = ["?????", "?????"];
     let teamIDs: Array<number> = [1, 2];
@@ -98,7 +87,6 @@ export default class Stats {
     return table;
   }
 
-
   private overallStatsTable(teamID: number, inGameID: number) {
     let table: HTMLTableElement = document.createElement("table");
     table.setAttribute("align", "center");
@@ -122,148 +110,6 @@ export default class Stats {
     }
 
     return table;
-  }
-
-  private matchViewer() {
-
-    let div = document.createElement("div");
-    div.id = "matchViewer";
-    div.style.fontFamily = "Tahoma, sans serif";
-
-    let title = document.createElement("b");
-    title.appendChild(document.createTextNode("Games"));
-
-    // Add buttons
-    let next = document.createElement("button");
-    next.setAttribute("class", "custom-button");
-    next.setAttribute("type", "button");
-    next.onclick = () => this.onNextMatch();
-    next.appendChild(this.images.controls.matchForward);
-
-    let back = document.createElement("button");
-    back.setAttribute("class", "custom-button");
-    back.setAttribute("type", "button");
-    back.onclick = () => this.onPreviousMatch();
-    back.appendChild(this.images.controls.matchBackward);
-
-    let gameNum = document.createElement("span");
-    gameNum.className += " gameNum";
-
-    title.appendChild(gameNum);
-    title.appendChild(back);
-    title.appendChild(next);
-    div.appendChild(title);
-    div.appendChild(document.createElement("br"));
-
-    return div;
-
-  }
-
-  refreshGameList(gameList: Array<Game>, activeGame: number, activeMatch: number) {
-
-    // Remove all games from the list
-    while(this.matches.childNodes[2]){
-      this.matches.removeChild(this.matches.childNodes[2]);
-    }
-
-    this.matches.childNodes[0].childNodes[1].textContent = " (" + (activeGame + 1) + "/" + gameList.length + ")";
-
-    //for (let game of gameList) {
-    for (var j = 0; j < gameList.length; j++) {
-      let game = gameList[j];
-      if(game != null) {
-
-        var metaData = game.meta;
-        var matchCount = game.matchCount;
-        var winner = game.winner;
-
-        // Construct a team vs. team string
-        var vsString = document.createElement("div");
-        var winnerString: HTMLSpanElement;
-        if(metaData != null) {
-
-          for (let team in metaData.teams) {
-              var teamName = document.createElement("span");
-              teamName.className += team === "1" ? " red" : " blue";
-              teamName.innerHTML = metaData.teams[team].name;
-              vsString.appendChild(teamName)
-              vsString.appendChild(document.createTextNode(" vs. "));
-              if(metaData.teams[team].teamID == winner) {
-                winnerString = teamName;
-              }
-          }
-
-          if(vsString.lastChild != null) {
-            vsString.removeChild(vsString.lastChild);
-          }
-
-          var gameDiv = document.createElement("div");
-          gameDiv.className += " gameDiv";
-          var title = document.createElement("b");
-          title.appendChild(vsString);
-          //title.appendChild(document.createTextNode(" on " + game.matchCount + " matches"))
-          // INSTEAD, Show i.e. Playing 1/3 << >>
-          gameDiv.appendChild(title);
-          if(game == gameList[activeGame]) {
-            gameDiv.appendChild(document.createTextNode("Playing match " + (activeMatch + 1) + "/" + matchCount));
-          }
-
-          gameDiv.appendChild(document.createElement("br"));
-
-          for (var i = 0; i < matchCount; i++) {
-            var match = game.getMatch(i);
-            var mapName = match.current.mapName;
-
-            var matchWinner = document.createElement("span");
-            for (let team in metaData.teams) {
-                if(metaData.teams[team].teamID == match.winner) {
-                  matchWinner.className += team === "1" ? " red" : " blue";
-                  matchWinner.innerHTML = metaData.teams[team].name;
-                  break;
-                }
-            }
-
-            // Add the information to the list
-            let matchEntry = document.createTextNode(" wins after " + (match.lastTurn + 1) + " rounds" );
-            let matchPrefix = document.createTextNode(mapName + " - ");
-            var matchWrapper = document.createElement("div");
-            matchWrapper.appendChild(matchPrefix);
-            matchWrapper.appendChild(matchWinner);
-            matchWrapper.appendChild(matchEntry);
-            matchWrapper.appendChild(document.createElement("br"));
-
-            if(j == activeGame && i == activeMatch) {
-              matchWrapper.className = 'active-match';
-            } else {
-              matchWrapper.className = 'inactive-match';
-
-              matchWrapper.onclick = (function(game, match, gotoMatch) {
-                  return function(){ gotoMatch(game, match); }
-              })(j,i, this.gotoMatch);
-
-            }
-
-            gameDiv.appendChild(matchWrapper);
-
-          }
-
-          // Create remove button
-          let remove = document.createElement("button");
-          remove.setAttribute("class", "custom-button");
-          remove.setAttribute("type", "button");
-          remove.textContent = "Remove";
-
-          remove.onclick = (function(game, removeGame) {
-                  return function(){ removeGame(game); }
-          })(j, this.removeGame);
-
-          gameDiv.appendChild(remove);
-
-          this.matches.appendChild(gameDiv);
-
-        }
-      }
-    }
   }
 
   /**
@@ -316,8 +162,6 @@ export default class Stats {
 
       this.div.appendChild(teamDiv);
     }
-
-    this.div.appendChild(this.matches);
   }
 
   /**
